+++ inst/include/epiworld/models/measlesmixing.hpp
@@ -65,10 +65,13 @@ using namespace epiworld;
  * - Hospitalized: Individuals requiring hospital care
  * - Recovered: Individuals who have recovered and gained immunity
  *
+ * ![Model Diagram](../assets/img/measlesmixing.png)
+ *
  * @tparam TSeq Type for genetic sequences (default: EPI_DEFAULT_TSEQ)
+ * @ingroup disease_specific
  */
 template<typename TSeq = EPI_DEFAULT_TSEQ>
-class ModelMeaslesMixing : public Model<TSeq>
+class ModelMeaslesMixing : public epiworld::Model<TSeq>
 {
 private:
 
@@ -776,6 +779,7 @@ inline void ModelMeaslesMixing<TSeq>::m_update_rash(
     }
     else if (which == 1) // Hospitalized
     {
+        m->record_hospitalization(*p);
         p->change_state(
             m,
             detected ?
@@ -840,6 +844,7 @@ inline void ModelMeaslesMixing<TSeq>::m_update_isolated(
     else if (which == 1)
     {
 
+        m->record_hospitalization(*p);
         if (unisolate)
         {
             p->change_state(