+++ inst/include/epiworld/models/measlesschool.hpp
@@ -44,9 +44,13 @@
  * detected. The agent is then isolated and all agents who are unvaccinated are
  * quarantined. Isolated agents then may be moved out of the isolation in
  * isolation_period days.
+ * 
+ * ![Model Diagram](../assets/img/measlesschool.png)
+ * 
+ * @ingroup disease_specific
  */
 template<typename TSeq = EPI_DEFAULT_TSEQ>
-class ModelMeaslesSchool: public Model<TSeq> {
+class ModelMeaslesSchool: public epiworld::Model<TSeq> {
 
 private:
 
@@ -76,6 +80,11 @@ private:
 
 public:
 
+    /**
+     * @name Model States
+     * @brief The different states of the model.
+     */
+    ///@{
     static constexpr epiworld_fast_uint SUSCEPTIBLE             = 0u;
     static constexpr epiworld_fast_uint EXPOSED                 = 1u;
     static constexpr epiworld_fast_uint PRODROMAL               = 2u;
@@ -89,6 +98,7 @@ public:
     static constexpr epiworld_fast_uint QUARANTINED_RECOVERED   = 10u;
     static constexpr epiworld_fast_uint HOSPITALIZED            = 11u;
     static constexpr epiworld_fast_uint RECOVERED               = 12u;
+    ///@}
     
     // Default constructor
     ModelMeaslesSchool() {};
@@ -498,6 +508,7 @@ LOCAL_UPDATE_FUN(m_update_rash) {
     {
         // If hospitalized, then the agent is removed from the system
         // effectively
+        model->record_hospitalization(*p);
         p->change_state(
             m,
             detected ?
@@ -557,6 +568,7 @@ LOCAL_UPDATE_FUN(m_update_isolated) {
     // If hospitalized, then the agent is removed from the system
     else if (which == 1u)
     {
+        model->record_hospitalization(*p);
         p->change_state(
             m,
             // HOSPITALIZED