+++ src/db.cpp
@@ -292,3 +292,92 @@ cpp11::writable::doubles get_today_total_cpp(SEXP model) {
   return totals_r;
 
 }
+
+[[cpp11::register]]
+cpp11::writable::data_frame get_active_cases_cpp(SEXP model) {
+
+  std::vector< int > date;
+  std::vector< int > virus;
+  std::vector< int > counts;
+
+  cpp11::external_pointer<Model<>> ptr(model);
+
+  ptr->get_db().
+    get_active_cases(date, virus, counts);
+
+  // Same idea with the names
+  std::vector< std::string > viruses;
+  for (const auto & i : ptr->get_viruses())
+    viruses.push_back(i->get_name());
+
+  std::vector< std::string > vnames(virus.size());
+  std::transform(
+    virus.begin(), virus.end(), vnames.begin(),
+    [&viruses](int i) { return viruses[i]; }
+  );
+
+  return cpp11::writable::data_frame({
+    "date"_nm      = date,
+    "virus_id"_nm  = virus,
+    "virus"_nm     = vnames,
+    "active_cases"_nm    = counts,
+  });
+
+}
+
+[[cpp11::register]]
+cpp11::writable::data_frame get_outbreak_size_cpp(SEXP model) {
+
+  std::vector< int > date;
+  std::vector< int > virus;
+  std::vector< int > counts;
+
+  cpp11::external_pointer<Model<>> ptr(model);
+
+  ptr->get_db().
+    get_outbreak_size(date, virus, counts);
+
+  // Same idea with the names
+  std::vector< std::string > viruses;
+  for (const auto & i : ptr->get_viruses())
+    viruses.push_back(i->get_name());
+
+  std::vector< std::string > vnames(virus.size());
+  std::transform(
+    virus.begin(), virus.end(), vnames.begin(),
+    [&viruses](int i) { return viruses[i]; }
+  );
+
+  return cpp11::writable::data_frame({
+    "date"_nm      = date,
+    "virus_id"_nm  = virus,
+    "virus"_nm     = vnames,
+    "outbreak_size"_nm    = counts,
+  });
+
+}
+
+
+[[cpp11::register]]
+cpp11::writable::data_frame get_hospitalizations_cpp(SEXP model) {
+
+  std::vector< int > date;
+  std::vector< int > virus_id;
+  std::vector< int > tool_id;
+  std::vector< int > count;
+  std::vector< double > weight;
+
+  cpp11::external_pointer<Model<>> ptr(model);
+
+  ptr->get_db().
+    get_hospitalizations(date, virus_id, tool_id, count, weight);
+
+  return cpp11::writable::data_frame({
+    "date"_nm     = date,
+    "virus_id"_nm = virus_id,
+    "tool_id"_nm  = tool_id,
+    "count"_nm    = count,
+    "weight"_nm   = weight
+  });
+
+}